backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic
python-multipart
passlib[bcrypt]
bcrypt<5.0.0
python-jose[cryptography]
